feat(MeshHTTPRoute): add header match to API
